requirements.txt
aiohttp
python-socketio
numpy